backend/requirements.txt
# FastAPI backend (phase 1). Python 3.10+
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
pyyaml>=6.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
